Update README.md
Hey there Folks!

This is our Capstone Project in which we are simulating a learning app which can help all the beginners out there who are struggling to get on roads for driving their own cars. This application can help them to achieve the very essential knowledge to clear the first and most significant hurdle to be eligible to drive their own vehicle on roads on Ontario,Canada.